Remove bit-smasher
The file delta.lisp used two functions from the bit-smasher library,
int->bits and bits->int. However this system was not specified as a
dependency for cl-git. Instead of adding a dependency, implement the
functions as they are straight-forward.

Note that bit-smasher 'reversed' the bit-vectors so we had to flip the
arrays to preserve that behavior. Because bit-vectors are arrays and
arrays are printed from low-to-high but octets are normally printed
from high-to-low (little-endian). If the bit-vectors were to use
big-endian instead the code for the helper could be simplified further.